refactor: fieldset 컴포넌트 분리 및 category 상수화
